Improve sitemap performance and prevent all sitemap pages to be updated when an old post changes
The performance improvement is mostly measured for individual sitemap pages for large
sites on pages with a high page number. The old query forted by date, which required
the full post table to be sorted before returning the right result.
The new solution removes the need for sorting and instead keeps a map of post start end end
ids of each page. This allows for fast lookups.
The creation of the map is still a slow process and will be improved in a next commit.

It used to be that, whenever an old post (on page 3) was updated, it would get bumped to the last page
(page 40) because of its now recent last_mod date. This would result in the oldest post of the next page (page 4)
to shift back one place, placing it on page 3, making its last_mod date more recent than the old value.
This in turn causes bots to crawl the sitemap again for no reason, as the updated content is not on that
sitemap page.